fix: CORUI-6264: One component per file Forms (#425)
* Rename file

* Move to temp

* Restore original

* Rename file

* Move to temp

* Restore original

* Rename file

* Move to temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename Forms to FormRadioItem

* Separate component files

* Missed index file
