+++ src/index.js
@@ -12,17 +12,17 @@ export { Calendar } from './Calendar/Calendar';
 export { ComboboxInput } from './ComboboxInput/ComboboxInput';
 export { DatePicker } from './DatePicker/DatePicker';
 export { Dropdown } from './Dropdown/Dropdown';
-export { FormSet } from './Forms/Forms';
-export { FormItem } from './Forms/Forms';
-export { FormLabel } from './Forms/Forms';
-export { FormInput } from './Forms/Forms';
-export { FormRadioGroup } from './Forms/Forms';
-export { FormRadioItem } from './Forms/Forms';
-export { FormTextarea } from './Forms/Forms';
-export { FormMessage } from './Forms/Forms';
-export { FormSelect } from './Forms/Forms';
-export { FormFieldset } from './Forms/Forms';
-export { FormLegend } from './Forms/Forms';
+export { default as FormSet } from './Forms/FormSet';
+export { default as FormItem } from './Forms/FormItem';
+export { default as FormLabel } from './Forms/FormLabel';
+export { default as FormInput } from './Forms/FormInput';
+export { default as FormRadioGroup } from './Forms/FormRadioGroup';
+export { default as FormRadioItem } from './Forms/FormRadioItem';
+export { default as FormTextarea } from './Forms/FormTextarea';
+export { default as FormMessage } from './Forms/FormMessage';
+export { default as FormSelect } from './Forms/FormSelect';
+export { default as FormFieldset } from './Forms/FormFieldset';
+export { default as FormLegend } from './Forms/FormLegend';
 export { Icon } from './Icon/Icon';
 export { Identifier } from './Identifier/Identifier';
 export { Image } from './Image/Image';
